virtualenv_localhost/requirements.txt
dogpile.core==0.4.1
ujson==1.35
wheel==0.26.0
